fix: harden the review findings and the newer toolchain's lints
The recogniser validates the buffer length against its dimensions
and refuses non-singleton model batches; the panel slicing panics on
a missing geometry field the way the original's lookup raises (the
scan worker contains it) instead of degrading to empty crops; the
region seams carry the window lookup's own integer width end to end;
the misleading score helper goes (nothing consumed it); the asset
exemption scopes match across both gates; the repair seam types are
named and the column means iterate, for the stricter lint.

+++ frontend/src-tauri/eo-services/src/skill_scan_manual.rs
@@ -39,7 +39,7 @@ pub const PAGE_COUNT: i64 = 12;
 pub const MAX_PAGE_COUNT: i64 = 30;
 
 /// One scan region: the panel's top-left and bottom-right corners.
-pub type ScanRegion = ([i32; 2], [i32; 2]);
+pub type ScanRegion = ([i64; 2], [i64; 2]);
 
 /// The capture observer (the recording controller's seam): called as
 /// `tap(panel, region, png)` after each successful page grab.
